+++ eng/ModVariantContentLoader/MainFile.cs
@@ -1,0 +1,9 @@
+using KitLib.ModVariantLoader;
+using MegaCrit.Sts2.Core.Modding;
+
+namespace KitLib.ModVariantLoader.ContentEntry;
+
+[ModInitializer(nameof(Initialize))]
+public static class MainFile {
+    public static void Initialize() => ModVariantBootstrap.Initialize();
+}